api/requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.115.2
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
psycopg2-binary==2.9.9
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.0.1
PyYAML==6.0.3
redis==5.0.8
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.30.6
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
